move the extension method back to static class
