+++ msgraph-sdk-raptor-compiler-lib/Models/RaptorConfig.cs
@@ -89,22 +89,5 @@ namespace MsGraphSDKSnippetsCompiler.Models
             get;
             init;
         }
-
-        /// <summary>
-        /// Extracts the configuration value, throws if empty string
-        /// </summary>
-        /// <param name="config">configuration</param>
-        /// <param name="key">lookup key</param>
-        /// <returns>non-empty configuration value if found</returns>
-        public static string GetNonEmptyValue(this IConfigurationRoot config, string key)
-        {
-            var value = config?.GetSection(key)?.Value;
-            if (string.IsNullOrWhiteSpace(value))
-            {
-                throw new InvalidDataException($"Value for {key} is not found in appsettings.json");
-            }
-
-            return value;
-        }
     }
 }
